Developed half of the page header
Added:
  - Fonts
  - Markup for logo, search field and catalog menu
  - Related sass components and basic styles
  - Related js components and basic utility functions

Main goal was to make use of html templates to change the content of
catalog menu and add styles to easily customize component's state.
Also it was important to provide keyboard access depending on whether
javascript is loaded or not. Javascript for search form and catalog menu
mostly deals with focus management and validation of search query.